Move isCountryInEu to own file
